Update order status

+++ src/main/java/com/fiap/techchallenge/adapters/in/rest/dto/CreateOrderDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class CreateOrderDTO {
 
     String costumerId;
 
-    Float amount;
+    BigDecimal amount;
 
 }
